refactor(tw): the ontology interfaces live in the types file
