+++ apps/tailwind-components/app/components/display/OntologyNode.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { ref } from "vue";
-import type { IOntologyTreeItem } from "../../utils/buildOntologyTree";
+import type { IOntologyTreeItem } from "../../../types/types";
 import CustomTooltip from "../CustomTooltip.vue";
 import BaseIcon from "../BaseIcon.vue";
